# bugfix attr-value judgment