--- a/slim_test.go
+++ b/slim_test.go
@@ -482,3 +482,42 @@ func TestAttr(t *testing.T) {
 		t.Fatalf("expected %v but %v", expect, got)
 	}
 }
+
+func TestIDAndClass(t *testing.T) {
+	tmpl, err := ParseFile("testdir/test_id_and_class.slim")
+	if err != nil {
+		t.Fatal(err)
+	}
+	var buf bytes.Buffer
+	err = tmpl.Execute(&buf, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	expect := readFile(t, "testdir/test_id_and_class.html")
+	got := buf.String()
+	if expect != got {
+		t.Fatalf("expected %v but %v", expect, got)
+	}
+}
+
+func TestIsAttributeValue(t *testing.T) {
+	tests := []struct {
+		in     rune
+		expect bool
+	}{
+		{'å', true},
+		{'A', true},
+		{'"', false},
+		{'\'', false},
+		{'=', false},
+		{'<', false},
+		{'>', false},
+		{'`', false},
+	}
+	for _, tt := range tests {
+		got := isUnquotedAttributeValue(tt.in)
+		if tt.expect != got {
+			t.Fatalf("expected %v but %v when in %s", tt.expect, got, string(tt.in))
+		}
+	}
+}