--- a/testdir/test_id_and_class.html
+++ b/testdir/test_id_and_class.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="ja">
+  <head>
+    <meta charset="UTF-8"/>
+    <title>your name.</title>
+  </head>
+  <body>
+    <div class="p-3 m-3">
+      <form id="form-box" class="form">
+        <label>Your Name</label>
+        <input type="text" name="your-name"/>
+        <input type="hidden" name="secret" value="1"/>
+      </form>
+    </div>
+  </body>
+</html>