--- a/testdir/test_id_and_class.slim
+++ b/testdir/test_id_and_class.slim
@@ -1,0 +1,11 @@
+doctype 5
+html lang="ja"
+  head
+    meta charset="UTF-8"
+    title your name.
+  body
+    div.p-3.m-3
+      form#form-box.form
+        label Your Name
+        input type=text name=your-name
+        input type=hidden name=secret value=1