I fixed the conditional expression of the attr-value.
But, in this PR isn't all patterns are met.

https://html.spec.whatwg.org/multipage/syntax.html#elements-2
```slim
div.p-3
```
```diff
- <div class="p">3</div>
+ <div class="p-3">
+ </div>
```

### benchmark
<details>
<summary>Old benchmarks</summary>

before::
```
BenchmarkTemplate_Execute-8   	  564622	      2096 ns/op
BenchmarkTemplate_Execute-8   	  565784	      2131 ns/op
BenchmarkTemplate_Execute-8   	  569425	      2124 ns/op
```
after::
```
BenchmarkTemplate_Execute-8   	  564313	      2117 ns/op
BenchmarkTemplate_Execute-8   	  563655	      2116 ns/op
BenchmarkTemplate_Execute-8   	  570637	      2118 ns/op
```

</details>

after merged benchmarks...
before::
```
BenchmarkTemplate_Execute-8   	  616350	      1914 ns/op
BenchmarkTemplate_Execute-8   	  615756	      1940 ns/op
BenchmarkTemplate_Execute-8   	  619611	      1922 ns/op
```
after::
```
BenchmarkTemplate_Execute-8   	  615942	      1937 ns/op
BenchmarkTemplate_Execute-8   	  616814	      1944 ns/op
BenchmarkTemplate_Execute-8   	  621721	      1913 ns/op
```